Pendings

+++ app/models/task.rb
@@ -19,9 +19,8 @@ class Task < ApplicationRecord
 
   validates_presence_of :creator_id, :owner_id, :property_id
   validates :priority, inclusion: { in: Constant::Task::PRIORITY, allow_blank: true, message: "must be one of these: #{Constant::Task::PRIORITY.to_sentence}" }
-  validates_inclusion_of  :license_required, :needs_more_info, :deleted, :hidden,
+  validates_inclusion_of  :license_required, :needs_more_info,
                           :initialization_template, in: [true, false]
-  validates_inclusion_of :status, in: %w[completed needsAction]
   validates_inclusion_of :visibility, in: [0, 1, 2, 3]
 
   validates :title, presence: true, uniqueness: { scope: :property }
@@ -31,14 +30,15 @@ class Task < ApplicationRecord
 
   monetize :budget_cents, :cost_cents, allow_nil: true
 
-  before_save :decide_completeness
-  before_save :sync_deleted_and_discarded_at, if: :unsynced_deleted_discard?
-  before_save :sync_completed_fields, if: -> { completed_at.present? || status == 'completed' }
+  before_save :decide_record_completeness
 
   after_create :create_with_api
   after_update :create_with_api, if: :saved_changes_to_users?
   after_update :update_with_api, if: :saved_changes_to_api_fields?
   after_update :relocate, if: -> { saved_change_to_property_id? }
+
+  # this probably isn't necessary because update_with_api should catch everything
+  # after_save :cascade_discarded, if: -> { discarded_at.present? }
 
   scope :needs_more_info, -> { undiscarded.where(needs_more_info: true).where(initialization_template: false) }
   scope :in_process, -> { undiscarded.where(completed_at: nil).where(initialization_template: false) }
@@ -64,25 +64,42 @@ class Task < ApplicationRecord
     return false if task_json.nil?
 
     tap do |t|
+      t.title = task_json['title']
       t.notes = task_json['notes']
-      t.status = task_json['status']
-      t.due = task_json['due']
-      t.deleted = task_json['deleted'] || false
-      t.hidden = task_json['hidden'] || false
       t.completed_at = task_json['completed']
+      t.due = task_json['due']
     end
 
     task_json.present?
   end
 
-  def create_taskuser_for(user)
-    tasklist = Tasklist.where(property: self.property, user: user).first
-    task_user = task_users.where(user: user).first_or_initialize
-    return 'already exists' if task_user.google_id.present?
-    response = TaskClient.new.insert(user, tasklist.google_id, self)
-    task_user.assign_from_api_fields!(response)
-    task_user.tasklist_id = tasklist.google_id
-    task_user.save
+  def create_taskuser_for(user, action = :insert)
+    tasklist = property.create_tasklist_for(user)
+    task_user = task_users.where(user: user).first_or_create
+    if task_user.google_id.nil?
+      response = TaskClient.new.send(
+        action,
+        user: user,
+        tasklist_gid: tasklist.google_id,
+        task: self,
+        task_user: task_user
+      )
+      task_user.assign_from_api_fields!(response)
+      task_user.update(tasklist_id: tasklist.google_id)
+      task_user.reload
+    end
+    task_user
+  end
+
+  def saved_changes_to_users?
+    saved_change_to_creator_id? || saved_change_to_owner_id?
+  end
+
+  def saved_changes_to_api_fields?
+    saved_change_to_title? ||
+      saved_change_to_notes? ||
+      saved_change_to_due? ||
+      saved_change_to_completed_at?
   end
 
   private
@@ -107,7 +124,7 @@ class Task < ApplicationRecord
     end
   end
 
-  def decide_completeness
+  def decide_record_completeness
     strikes = 0
 
     strikes += 3 if due.nil?
@@ -118,60 +135,33 @@ class Task < ApplicationRecord
     true
   end
 
-  def unsynced_deleted_discard?
-    return false if !deleted? && discarded_at.blank?
-    return false if deleted? && discarded_at.present?
-    true
-  end
-
-  def sync_deleted_and_discarded_at
-    self.discarded_at = Time.now if deleted?
-    self.deleted = discarded_at.present? ? true : false
-  end
-
-  def sync_completed_fields
-    return true if completed_at.present? && status == 'completed'
-    self.completed_at = Time.now
-    self.status = 'completed'
-  end
-
-  def saved_changes_to_api_fields?
-    saved_change_to_title? ||
-      saved_change_to_notes? ||
-      saved_change_to_due? ||
-      saved_change_to_status? ||
-      saved_change_to_deleted? ||
-      saved_change_to_completed_at?
-  end
-
-  def saved_changes_to_users?
-    saved_change_to_creator_id? || saved_change_to_owner_id?
+  def prepare_for_api
+    [creator, owner].each do |user|
+      create_taskuser_for(user)
+      Property.find(property_id_before_last_save).create_tasklist_for(user) if property_id_before_last_save.present?
+    end
   end
 
   def create_with_api
-    [creator, owner].each do |user|
-      property.create_tasklist_for(user)
-    end
-
-    [owner, creator].each do |user|
-      create_taskuser_for(user)
-    end
+    prepare_for_api
   end
 
   def update_with_api
-    [creator, owner].each do |user|
-      property.create_tasklist_for(user)
-      Property.find(property_id_before_last_save).create_tasklist_for(user)
-      create_taskuser_for(user)
-    end
-
+    prepare_for_api
     action = discarded_at.present? ? :delete : :update
 
-    [owner, creator].each do |user|
-      tasklist = Tasklist.where(property: property, user: user).first
+    [creator, owner].each do |user|
       task_user = task_users.where(user: user).first
-      response = TaskClient.new.send(action, user, task_user.tasklist_id, self, task_user.google_id) if task_user.tasklist_id.present?
-      if action == :delete
+      response = TaskClient.new.send(
+        action,
+        user: user,
+        task: self,
+        task_user: task_user,
+        tasklist_gid: task_user.tasklist_id,
+        task_gid: task_user.google_id
+      )
+
+      if discarded_at.present?
         task_user.destroy
       else
         task_user.assign_from_api_fields!(response)
@@ -179,27 +169,24 @@ class Task < ApplicationRecord
       end
     end
 
-    if saved_change_to_creator_id?
-      task_users.where(user_id: creator_id_before_last_save).first.destroy
-    end
-
-    if saved_change_to_owner_id?
-      task_users.where(user_id: owner_id_before_last_save).frst.destroy
-    end
+    task_users.where(user_id: creator_id_before_last_save).first.destroy if saved_change_to_creator_id?
+    task_users.where(user_id: owner_id_before_last_save).frst.destroy    if saved_change_to_owner_id?
   end
 
   def relocate
+    prepare_for_api
     [creator, owner].each do |user|
-      property.create_tasklist_for(user)
-      Property.find(property_id_before_last_save).create_tasklist_for(user)
-      create_taskuser_for(user)
-    end
-
-    [owner, creator].each do |user|
       old_tasklist = Tasklist.where(property_id: property_id_before_last_save, user: user).first
       tasklist = Tasklist.where(property: property, user: user).first
       task_user = task_users.where(user: user).first
-      TaskClient.new.relocate(user, old_tasklist.google_id, tasklist.google_id, self, task_user.google_id) if task_user.tasklist_id.present?
+      TaskClient.new.relocate(
+        user: user,
+        old_list_gid: old_tasklist.google_id,
+        new_list_gid: tasklist.google_id,
+        task: self,
+        task_user: task_user,
+        task_gid: task_user.google_id
+      )
     end
   end
 end
